% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_add_element.R
\name{dist_add_elements}
\alias{dist_add_elements}
\alias{dist_add_curve}
\alias{dist_add_shading}
\alias{dist_add_segment}
\title{Add Distribution Elements to a Plot}
\usage{
dist_add_elements(
  data = NULL,
  linetype = "solid",
  linewidth = 1,
  color = "black",
  shade_type = NULL,
  shade_limits = NULL,
  shade_fill = "red",
  shade_alpha = 0.7,
  seg_x = NULL,
  seg_linetype = "solid",
  seg_linewidth = 1,
  seg_color = "black",
  ...
)

dist_add_curve(
  data = NULL,
  linetype = "solid",
  linewidth = 1,
  color = "black",
  ...
)

dist_add_shading(
  mapping = ggplot2::aes(x = x, y = y),
  data = NULL,
  shade_type = c("above", "below", "between", "tails"),
  shade_limits = NULL,
  shade_fill = "red",
  shade_alpha = 0.7,
  ...
)

dist_add_segment(
  mapping = ggplot2::aes(x = x, y = y),
  data = NULL,
  seg_x,
  seg_linetype = "solid",
  seg_linewidth = 1,
  seg_color = "black",
  ...
)
}
\arguments{
\item{data}{Optional data frame containing columns named \code{x} and \code{y} (e.g.,
created with one of the \verb{gen_*_tbl()} functions, such as
\code{gen_norm_dist()}). The \code{x}column contains the full range of values across
the x-axis, and the \code{y} column contains the density value for that value of
\code{x}. If not provided, will use the data already associated with the plot
object.}

\item{linetype, linewidth, color}{Parameters passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} to
draw the curve (defaults to a solid black line).}

\item{shade_type}{Character scalar. Type of shading to apply to a specific
area under the curve. Options are \code{below}, \code{above}, \code{between}, \code{tails}. If
no shading is desired, set to \code{NULL} (the default).}

\item{shade_limits}{Numeric. When applying shading, specifies the lower/upper
limit of the shading to be applied. Use a single value when \code{shade_type} is
set to \code{below} or \code{above}, and a vector of length 2 when \code{shade_type} is
set to \code{between} or \code{tails}.}

\item{shade_fill, shade_alpha}{Parameters passed to \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_area()}} to
shade the area under the curve. (defaults to red with some translucency).}

\item{seg_x}{Numeric vector of \code{x} values where the segments should be drawn.
The segments will go from zero (x-axis) to the height of the density curve.}

\item{seg_linetype, seg_linewidth, seg_color}{Parameters passed to
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} to draw the line segments (defaults to a solid black
line).}

\item{...}{Additional arguments passed to \verb{[ggplot2::ggplot2()]} geoms
(\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}, \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_area()}},
\code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}).}

\item{mapping}{\verb{[ggplot2::aes()]} mapping of parameters (defaults to \code{ggplot2::aes(x = x, y = y)})}
}
\value{
ggplot2 object.
}
\description{
Workhorse functions to add individual elements to an existing plot. These are
generally intended for internal use by the higher level functions (e.g.,
\code{plot_dist()}, \code{add_dist()}, etc.), but can be useful to users in some contexts.
}
\details{
\code{dist_add_elements()} will optionally add the distribution curve, shading,
and vertical line segments at once, in an order that ensures that the curve
does not get overlapped by the shading or line segments. The other functions
add one individual element (curve: \code{dist_add_curve()}, shading:
\code{dist_add_shading()}, vertical line segment: \code{dist_add_segment()}), allowing
for more flexibility or the possibility of multiple layers..
}
\examples{
# Generate some data and (relatively) blank plot:
norm.tbl <- gen_normal_tbl(M = 0, SD = 1)
p <- plot_dist(data = norm.tbl)

p + dist_add_elements(data = gen_normal_tbl(M = 1),
											linetype = 'dotted', shade_type = 'above', shade_limits = 1,
											seg_x = 1, seg_linetype = 'dotted')

p + dist_add_curve(data = gen_normal_tbl(M = 2))

p + dist_add_shading(shade_type = 'below', shade_limits = 1, shade_fill = "red")

p + dist_add_segment(seg_x = -2:2)

p +
	dist_add_shading(shade_type = 'below', shade_limits = -2, shade_fill = 'blue') +
	dist_add_shading(shade_type = 'between', shade_limits = c(-2,-1), shade_fill = 'red') +
	dist_add_shading(shade_type = 'between', shade_limits = c(-1,0), shade_fill = 'blue') +
	dist_add_shading(shade_type = 'between', shade_limits = c(0,1), shade_fill = 'red') +
	dist_add_shading(shade_type = 'between', shade_limits = c(1,2), shade_fill = 'blue') +
	dist_add_shading(shade_type = 'above', shade_limits = 2, shade_fill = 'red') +
	dist_add_segment(seg_x = -2:2)

}
