% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistPlot-package.R
\docType{package}
\name{DistPlot-package}
\alias{DistPlot}
\alias{DistPlot-package}
\title{DistPlot: Create an interactive density plot}
\description{
This package creates 2 plots representing the Normal density and the Student density. There are interactive.
}
\author{
\strong{Maintainer}: Léa Jacquet \email{lea.jacquet@student.uclouvain.be}

}
\keyword{internal}
