% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distplot.R
\docType{package}
\name{distplot}
\alias{distplot-package}
\alias{distplot}
\title{distplot: A package for generating statistical distributions with simple,
minimalist aesthetics.}
\description{
The distplot package has functions to generate several types of distributions
based on provided parameters (e.g., normal, t, and F distributions), with
optional shading and added line segments.
}
\author{
\strong{Maintainer}: Travis Carter \email{tjcarter@roanoke.edu} (\href{https://orcid.org/0000-0003-4499-7300}{ORCID})

}
