% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{distribution_scaling}
\alias{distribution_scaling}
\title{Distribution Scale Aesthetics}
\usage{
distribution_scaling(
  x_min,
  x_max,
  x_breaks = seq(x_min, x_max, by = 1),
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{x_min}{Numeric scalar. Lower limit of the x-axis scale.}

\item{x_max}{Numeric scalar. Upper limit of the x-axis scale.}

\item{x_breaks}{Numeric vector. Vector of values to use for the x-axis. If
not provided, inferred to be integers ranging from x_min to x_max. Set to
\code{NULL} to remove values.}

\item{xlab}{Character scalar. Title of x-axis.}

\item{ylab}{Character scalar. Title of y-axis. (Note: if using
\code{\link[=distribution_thm]{distribution_thm()}}, setting \code{y_line} set to \code{FALSE} will suppress the
y-axis title regardless of the value of \code{ylab} set here.)}
}
\description{
Applies sensible defaults to the properties of the x- and y-axis scales when
plotting distributions, given range of x. Apply values by adding function to
an existing ggplot object (see examples). Primarily used internally as part
of the plotting functions (e.g., \code{\link[=plot_norm]{plot_norm()}}).
}
\examples{

df <- tibble::tibble(x = seq(-4, 4, length.out = 1001), y = dnorm(x))
p <- ggplot2::ggplot(df, ggplot2::aes(x = x, y = y)) +
     ggplot2::geom_line() + distribution_thm()
p + distribution_scaling(x_min = -3.5, x_max = 3.5, x_breaks = -3:3)

p + distribution_scaling(x_min = -4, x_max = 4, xlab = 'Z')

}
