% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_plot.R
\name{distribution_thm}
\alias{distribution_thm}
\title{ggplot Theme for Distributions}
\usage{
distribution_thm(
  x_ticks = TRUE,
  x_line = TRUE,
  y_ticks = FALSE,
  y_line = FALSE,
  ...
)
}
\arguments{
\item{x_ticks}{Logical. Should the plot include ticks on the x-axis?}

\item{x_line}{Logical. Should the plot show the horizontal line of the
x-axis? (If set to \code{FALSE}, recommend setting \code{x_ticks} to \code{FALSE} as
well.)}

\item{y_ticks}{Logical. Should the plot include ticks on the y-axis?}

\item{y_line}{Logical. Should the plot show the vertical line of the y-axis?
(If set to \code{FALSE}, recommend setting \code{x_ticks} to \code{FALSE} as well.)}

\item{...}{Additional parameters passed to \code{\link[ggplot2:ggtheme]{ggplot2::theme_classic()}}.}
}
\value{

}
\description{
Bare bones theme (based on \code{\link[ggplot2:ggtheme]{ggplot2::theme_classic()}}) for plotting
distributions with sensible defaults. Makes it easy to add/remove the ticks
and line for the x- and y-axes. For plotting functions this is automatically
applied, along with other tweaks to the x- and y-axis scales. Apply theme by
adding function to an existing ggplot object (see examples).
}
\examples{

p <- tibble::tibble(x = seq(-4, 4, length.out = 1001), y = dnorm(x)) \%>\%
ggplot2::ggplot(ggplot2::aes(x = x, y = y)) +
ggplot2::geom_line()
p + distribution_thm()

p + distribution_thm(y_line = TRUE, y_ticks = TRUE)

}
