% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t.R
\name{gen_t_tbl}
\alias{gen_t_tbl}
\title{Generate Student's \emph{t} Distribution Data}
\usage{
gen_t_tbl(df, x_min = -4, x_max = 4, padding = 0.1, n = 1001)
}
\arguments{
\item{df}{Numeric scalar. Degrees of freedom for the \emph{t} distribution.}

\item{x_min}{Numeric scalar. Lower limit of the x-axis scale.}

\item{x_max}{Numeric scalar. Upper limit of the x-axis scale.}

\item{padding}{Numeric scalar. How much padding should be added to the range
of \code{x} (expressed as proportion)? Defaults to 10\% (0.1).}

\item{n}{Numeric scalar. Length of sequence for the range of \code{x} values to
generate.}
}
\value{

}
\description{
Given degrees of freedom, generates a \code{tibble} with a range of values of \code{x}
and their corresponding \emph{t}-distribution density values as \code{y}. This
function is useful for generating distribution plots.
}
\examples{
gen_t_tbl(df = 100)
}
