% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlotDens.R
\name{runPlotDens}
\alias{runPlotDens}
\title{runPlotDens Function}
\usage{
runPlotDens()
}
\value{
Interactive Shiny App
}
\description{
This function will open an interactive Shiny App where the user can choose many parameters to plot the normal and/or Student-t Distributions.
}
