% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDens.R
\name{PlotDens}
\alias{PlotDens}
\title{PlotDens Function}
\usage{
PlotDens(a = -1, b = 1, mu = 0, var = 1, df = 1, output = 1)
}
\arguments{
\item{a}{Starting of the density area}

\item{b}{End of the density area}

\item{mu}{Mean of the normal density}

\item{var}{Variance of the normal density}

\item{df}{Degree og freedom of the student density}

\item{output}{The plot you want to disply (1 = both, 2= only normal distribution, 3= Only student distribution)}
}
\value{

}
\description{
Allow beginner students in statistic to visualize and compare a normal and student distribution.
}
\examples{
PlotDens( a = 2, b=2, mu=1, var=1.7, df=4, output=1)

}
