% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{PlotDens}
\alias{PlotDens}
\title{Normal vs the Student}
\usage{
PlotDens(mean = 0, sd = 1, df = 1, a = -1, b = 1, output = "overlay")
}
\arguments{
\item{mean}{mean of the Normal distribution}

\item{sd}{Standard Deviation of the Normal distribution}

\item{df}{Degree of freedom of the Student's t distribution}

\item{a}{lower bound of the interval of the shaded area}

\item{b}{upper bound of the interval of the shaded area}

\item{output}{which density to plot? "norm", "t", or "overlay" (both). This latter is the default.}
}
\value{
ggplot object
}
\description{
Plot both the Normal density and the Student density using ggplot. The plot also shows shaded areas under each density corresponding to a user-specified interval.
}
\examples{
PlotDens()

}
