% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDens.R
\name{PlotDens}
\alias{PlotDens}
\title{Plots of densities}
\usage{
PlotDens(a = -1, b = 1, mean = 0, sd = 1, df = 1, out = 1)
}
\arguments{
\item{a}{is an integer between \link{-5:5}}

\item{b}{is an integer between \link{-5:5}}

\item{mean}{give the mean for the normal density}

\item{sd}{give the standard deviation for the normal density}

\item{df}{give the degree of freedom of the student density}

\item{out}{is an integer that select the choosen density plot (1= both, 2=normal density, 3=student)}
}
\value{
density plots
}
\description{
The function takes 4 arguments and creates 2 plots (a Normal density plot and a Student density plot)
}
\examples{
PlotDens (a=-1, b=1, mean=0, sd=1, df= 1, out=1)

}
