% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{runPlotDens}
\alias{runPlotDens}
\title{Normal vs the Student}
\usage{
runPlotDens()
}
\value{
shiny app
}
\description{
Plot both the Normal density and the Student density \strong{interactively} using shiny.
}
\examples{
# runPlotDens()

}
